Split video player managers
