+++ Shared/ViewModels/VideoPlayerManager/VideoPlayerManager.swift
@@ -276,61 +276,7 @@ class VideoPlayerManager: ViewModel {
             let request = Paths.reportPlaybackProgress(progressInfo)
             let _ = try await userSession.client.send(request)
 
-            print("sent progress task")
+            logger.trace("sent progress task")
         }
     }
-}
-
-// TODO: move to own file
-class OnlineVideoPlayerManager: VideoPlayerManager {
-
-    init(item: BaseItemDto, mediaSource: MediaSourceInfo) {
-        super.init()
-
-        Task {
-            let viewModel = try await item.videoPlayerViewModel(with: mediaSource)
-
-            await MainActor.run {
-                self.currentViewModel = viewModel
-            }
-        }
-    }
-}
-
-// TODO: move to own file
-class DownloadVideoPlayerManager: VideoPlayerManager {
-
-    init(downloadTask: DownloadTask) {
-        super.init()
-
-        guard let playbackURL = downloadTask.getMediaURL() else {
-            logger.error("Download task does not have media url for item: \(downloadTask.item.displayTitle)")
-
-            return
-        }
-
-        self.currentViewModel = .init(
-            playbackURL: playbackURL,
-            item: downloadTask.item,
-            mediaSource: .init(),
-            playSessionID: "",
-            videoStreams: downloadTask.item.videoStreams,
-            audioStreams: downloadTask.item.audioStreams,
-            subtitleStreams: downloadTask.item.subtitleStreams,
-            selectedAudioStreamIndex: 1,
-            selectedSubtitleStreamIndex: 1,
-            chapters: downloadTask.item.fullChapterInfo,
-            streamType: .direct
-        )
-    }
-
-    override func getAdjacentEpisodes(for item: BaseItemDto) {}
-
-    override func sendStartReport() {}
-
-    override func sendPauseReport() {}
-
-    override func sendStopReport() {}
-
-    override func sendProgressReport() {}
 }
